+++ lib/src/scm.dart
@@ -356,6 +356,7 @@ class Scm implements ScmNodeInterface {
         node.productionStartTime = _stopwatch.elapsed;
 
         // Produce
+        print('Producing node: $node');
         node.produce();
       }