+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+
+- Fix issues in supply chain
+
 ## [1.0.4] - 2024-05-08
 
 ### Changed
@@ -50,6 +56,7 @@
 - 'Github Actions Pipeline'
 - 'Github Actions Pipeline: Add SDK file containing flutter into .github/workflows to make github installing flutter and not dart SDK'
 
+[Unreleased]: https://github.com/inlavigo/supply_chain/compare/1.0.4...HEAD
 [1.0.4]: https://github.com/inlavigo/supply_chain/compare/1.0.3...1.0.4
 [1.0.3]: https://github.com/inlavigo/supply_chain/compare/1.0.2...1.0.3
 [1.0.2]: https://github.com/inlavigo/supply_chain/compare/1.0.1...1.0.2